refactor: repository 인터페이스 도메인 모듈로 이동
