+++ baseball/src/main/java/com/flab/baseball/application/GameHistoriesSearchProcessor.java
@@ -2,7 +2,7 @@ package com.flab.baseball.application;
 
 import com.flab.baseball.application.data.GameHistoriesData;
 import com.flab.baseball.domain.GameHistory;
-import com.flab.baseball.infrastructure.GameHistoryRepositoryAdaptor;
+import com.flab.baseball.domain.repository.GameHistoryRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -11,10 +11,10 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class GameHistoriesSearchProcessor {
 
-	private final GameHistoryRepositoryAdaptor gameHistoryRepositoryAdaptor;
+	private final GameHistoryRepository gameHistoryRepository;
 
 	public GameHistoriesData findHistories(Long roomId) {
-		List<GameHistory> gameHistories = gameHistoryRepositoryAdaptor.findAllByRoomId(roomId);
+		List<GameHistory> gameHistories = gameHistoryRepository.findAllByRoomId(roomId);
 		return new GameHistoriesData(gameHistories);
 	}
