+++ baseball/src/main/java/com/flab/baseball/application/RoomResultProcessor.java
@@ -1,8 +1,8 @@
 package com.flab.baseball.application;
 
-import com.flab.baseball.domain.Room;
-import com.flab.baseball.infrastructure.RoomRepositoryAdaptor;
 import com.flab.baseball.application.data.RoomResultData;
+import com.flab.baseball.domain.Room;
+import com.flab.baseball.domain.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -10,10 +10,10 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class RoomResultProcessor {
 
-	private final RoomRepositoryAdaptor roomRepositoryAdaptor;
+	private final RoomRepository roomRepository;
 
 	public RoomResultData roomResult(Long roomId) {
-		Room room = roomRepositoryAdaptor.getRoomById(roomId);
+		Room room = roomRepository.findRoomById(roomId);
 		return new RoomResultData(room.getRound());
 	}
