+++ baseball/src/main/java/com/flab/baseball/infrastructure/GameHistoryRepositoryAdaptor.java
@@ -1,15 +1,16 @@
 package com.flab.baseball.infrastructure;
 
 import com.flab.baseball.domain.GameHistory;
+import com.flab.baseball.domain.repository.GameHistoryRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class GameHistoryRepositoryAdaptor {
+public class GameHistoryRepositoryAdaptor implements GameHistoryRepository {
 
-	private final GameHistoryRepository gameHistoryRepository;
+	private final MemoryGameHistoryRepository gameHistoryRepository;
 
 	public GameHistory persist(GameHistory gameHistory) {
 		return gameHistoryRepository.persist(gameHistory);
